% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_erm.R
\name{read_erm}
\alias{read_erm}
\title{Read eRm Objects for Plink Analysis}
\usage{
read_erm(x, loc.out = FALSE, as.irt.pars = TRUE)
}
\arguments{
\item{x}{An eRm model object (RM, RSM, or PCM)}

\item{loc.out}{Logical. Should location parameters be output? Default is FALSE.}

\item{as.irt.pars}{Logical. Should the output be returned as irt.pars object? Default is TRUE.}
}
\value{
If as.irt.pars = TRUE, returns an irt.pars object suitable for plink functions.
If as.irt.pars = FALSE, returns a list of parameter matrices.
}
\description{
A modified version of plink::read.erm that actually works correctly with eRm objects.
This function converts eRm model objects to the format required by the plink package
for test linking and equating procedures.
}
\details{
This function addresses issues with the original plink::read.erm function
by properly setting the groups and time variables to 1, which are not part of the
standard output from eRm objects. The function:
\itemize{
\item Extracts item parameters from eRm objects
\item Converts them to the plink format
\item Handles dichotomous (RM) and polytomous (RSM, PCM) models
\item Returns parameters in irt.pars format for linking
}
}
\examples{
\dontrun{
# Convert eRm model for plink analysis
plink_pars <- read_erm(rasch_model)

# Get raw parameter matrices
raw_pars <- read_erm(rasch_model, as.irt.pars = FALSE)
}
}
